requirements.txt
alembic==1.6.5
aniso8601==9.0.1
attrs==21.2.0
bidict==0.21.2
cassandra-driver==3.25.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dnspython==1.16.0
eventlet==0.31.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Migrate==2.5.3
Flask-RESTful==0.3.9
flask-restx==0.3.0
Flask-SocketIO==3.0.1
Flask-SQLAlchemy==2.4.1
Flask-CQLAlchemy==1.2.0
gevent==21.1.2
greenlet==1.1.0
idna==2.10
importlib-metadata==4.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==2.0.1
PyJWT==2.0.1
PyMySQL==1.0.2
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.13.2
python-socketio==4.3.1
pytz==2021.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.3.15
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
